Complete framework
Add `pyproject.toml` with project metadata, dependencies, and build system configuration.

**Refactor and update imports:**
* Update imports in `main.py`, `examples.py`, `llm/__init__.py`, and `llm/providers/openai.py` to reflect new file structure.
* Remove unused imports and refactor code for better readability.

**Add and update type hints:**
* Add type hints in `main.py`, `examples.py`, `llm/base.py`, `llm/config.py`, and `llm/providers/anthropic.py` to make the project type strict.

**Remove unused code:**
* Delete `models/message.py`.

**Add new files for better organization:**
* Add `models/personality.py`, `models/demographics.py`, and `models/memory.py` to separate model definitions.

**Update README.md:**
* Update documentation to reflect new file structure and provide installation and usage instructions.
