+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "agient"
+version = "0.1.0"
+description = "A Psychologically-Grounded Agent Simulation Framework"
+authors = ["Daniel Merja <daniel.merja@example.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.11"
+pydantic = "^1.8"
+openai = "^0.27.0"
+anthropic = "^0.2.0"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.2"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
