+++ tests/test_storage.py
@@ -1,4 +1,3 @@
-
 import unittest
 from uuid import uuid4
 from datetime import datetime, timedelta
